+++ lib/common/total.dart
@@ -1,17 +1,19 @@
+import 'package:decimal/decimal.dart';
+
 class Total {
   DateTime timeStamp;
-  double volumeUSD;
-  double liquidityUSD;
+  Decimal volumeUSD;
+  Decimal liquidityUSD;
 
-  Total({DateTime timeStamp, double volumeUSD, double liquidityUSD})
+  Total({DateTime timeStamp, Decimal volumeUSD, Decimal liquidityUSD})
       : this.timeStamp = timeStamp,
         this.volumeUSD = volumeUSD,
         this.liquidityUSD = liquidityUSD;
 
   Total.fromJson(Map<String, dynamic> json)
       : timeStamp = DateTime.parse(json['Time']),
-        volumeUSD = double.parse(json['volumeUSD']),
-        liquidityUSD = double.parse(json['liquidityUSD']);
+        volumeUSD = Decimal.parse(json['volumeUSD']),
+        liquidityUSD = Decimal.parse(json['liquidityUSD']);
 
   Map<String, dynamic> toJson() =>
       {'Time': timeStamp, 'volumeUSD': volumeUSD, 'liquidityUSD': liquidityUSD};
